+++ webcomponents/inline-editor/src/components/actions/list-actions/list-actions.tsx
@@ -1,8 +1,7 @@
 import {Component, Event, EventEmitter, h, Host, Prop} from '@stencil/core';
 
-import {DeckdeckgoInlineEditorUtils} from '../../../utils/utils';
-
 import {ContentList} from '../../../types/enums';
+import {ExecCommandAction} from '../../../interfaces/interfaces';
 
 @Component({
   tag: 'deckgo-ie-list-actions',
@@ -26,15 +25,18 @@ export class AlignActions {
   contentList: ContentList;
 
   @Event()
-  private listModified: EventEmitter<void>;
+  private execCommand: EventEmitter<ExecCommandAction>;
 
-  private toggleList(e: UIEvent, cmd: string): Promise<void> {
+  private toggleList(e: UIEvent, type: 'ol' | 'ul'): Promise<void> {
     return new Promise<void>(async (resolve) => {
       e.stopPropagation();
 
-      await DeckdeckgoInlineEditorUtils.execCommand(this.selection, cmd);
-
-      await this.listModified.emit();
+      this.execCommand.emit({
+        cmd: 'list',
+        detail: {
+          type,
+        },
+      });
 
       resolve();
     });
@@ -46,7 +48,7 @@ export class AlignActions {
         <deckgo-ie-action-button
           mobile={this.mobile}
           disableAction={this.disabledTitle}
-          onAction={($event: CustomEvent<UIEvent>) => this.toggleList($event.detail, 'insertOrderedList')}
+          onAction={($event: CustomEvent<UIEvent>) => this.toggleList($event.detail, 'ol')}
           class={this.contentList === ContentList.ORDERED ? 'active' : undefined}>
           <deckgo-ie-action-image cssClass={'ordered-list'}></deckgo-ie-action-image>
         </deckgo-ie-action-button>
@@ -54,7 +56,7 @@ export class AlignActions {
         <deckgo-ie-action-button
           mobile={this.mobile}
           disableAction={this.disabledTitle}
-          onAction={($event: CustomEvent<UIEvent>) => this.toggleList($event.detail, 'insertUnorderedList')}
+          onAction={($event: CustomEvent<UIEvent>) => this.toggleList($event.detail, 'ul')}
           class={this.contentList === ContentList.UNORDERED ? 'active' : undefined}>
           <deckgo-ie-action-image cssClass={'unordered-list'}></deckgo-ie-action-image>
         </deckgo-ie-action-button>