+++ webcomponents/inline-editor/src/interfaces/interfaces.ts
@@ -13,3 +13,18 @@ export interface InlineAction {
   selection: Selection;
   anchorLink: AnchorLink;
 }
+
+export interface ExecCommandStyle {
+  style: 'color' | 'background-color' | 'font-size' | 'font-weight' | 'font-style' | 'text-decoration';
+  value: string;
+  initial: (element: HTMLElement | null) => Promise<boolean>;
+}
+
+export interface ExecCommandList {
+  type: 'ol' | 'ul';
+}
+
+export interface ExecCommandAction {
+  cmd: 'style' | 'list';
+  detail: ExecCommandStyle | ExecCommandList;
+}